% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{create.DOT}
\alias{create.DOT}
\title{Creating a DOT object based on the processed ref and srt data}
\usage{
create.DOT(srt, ref, ls_solution = TRUE)
}
\arguments{
\item{srt}{A list containing the processed srt data produced by setup.srt}

\item{ref}{A list containing the processed ref data produced by setup.ref}

\item{ls_solution}{Boolean. Whether an initial solution based on LS should be produced}
}
\value{
A DOT object ready to be fed to the algorithm
}
\description{
Creating a DOT object based on the processed ref and srt data
}
\examples{
data(dot.sample)
dot.ref <- setup.ref(dot.sample$ref$counts[, 1:1000], dot.sample$ref$labels[1:1000], 2)
dot.srt <- setup.srt(dot.sample$srt$counts, dot.sample$srt$coordinates)
dot <- create.DOT(dot.srt, dot.ref)
}
