% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\docType{class}
\name{Dot-class}
\alias{Dot-class}
\alias{Dot}
\title{An S4 class to represent a DOT object.}
\description{
An S4 class to represent a DOT object.
}
\section{Slots}{

\describe{
\item{\code{srt}}{A list containing spatial data as processed by setup.srt}

\item{\code{ref}}{A list containing reference data as processed by setup.ref}

\item{\code{weights}}{A matrix containing absolute abundance of categories}

\item{\code{solution}}{A matrix containing raw solution}

\item{\code{history}}{A data.frame containing solution history}
}}

