% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT-package.R
\docType{package}
\name{DOT-package}
\alias{DOT}
\alias{DOT-package}
\title{DOT: DOT: A flexible multi-objective optimization framework for transferring features across single-cell and spatial omics}
\description{
DOT is a method for transferring cell features from a reference single-cell RNA-seq data to spots/cells in spatial omics. It operates by optimizing a combination of multiple objectives using a Frank-Wolfe algorithm to produce a high quality transfer. Apart from transferring cell types/states to spatial omics, DOT can be used for transferring other relevant categorical or continuous phenotypes from one set of omics to another, such as estimating the expression of missinng genes or transferring transcription factor/pathway activities.
}
\author{
\strong{Maintainer}: Arezou Rahimi \email{arezou.rahimi@uni-heidelberg.de} (\href{https://orcid.org/0000-0003-0285-8089}{ORCID})

}
\keyword{internal}
