% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dot.sample}
\alias{dot.sample}
\title{Sample reference scRNA-seq and spatial data}
\format{
A \code{"list"}
}
\usage{
data(dot.sample)
}
\description{
Sample reference scRNA-seq and synthetic multicell spatial data of the primary motor cortex region (MOp) of the mouse brain.
The reference single-cell RNA seq sample comes from a similar region and contains 44 cell types.
Both spatial and single-cell data are down-sampled to 500 genes in this example.
}
\keyword{datasets}
