% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{draw_maps}
\alias{draw_maps}
\title{A plotting wrapper for drawing gene/abundance maps on tissue}
\usage{
draw_maps(
  spatial,
  weights,
  normalize = TRUE,
  ncol = 4,
  trans = "identity",
  size = 0.5,
  shape = 19,
  flip_y = TRUE,
  viridis_option = "magma",
  legend_title = "abundance",
  background = "gray90"
)
}
\arguments{
\item{spatial}{Coordinates}

\item{weights}{Weights to draw (spot x feature)}

\item{normalize}{Whether weights should sum up to 1 for each spot}

\item{ncol}{Number of columns in the grid}

\item{trans}{Color transformation scale}

\item{size}{Size of points}

\item{shape}{Shape of points}

\item{flip_y}{Whether y axis should be negated}

\item{viridis_option}{Viridis color options}

\item{legend_title}{Title of color legend}

\item{background}{Background color of plot}
}
\value{
A ggplot object
}
\description{
A plotting wrapper for drawing gene/abundance maps on tissue
}
\examples{
data(dot.sample)
draw_maps(dot.sample$srt$coordinates, t(as.matrix(dot.sample$srt$counts[1:8, ])), normalize = FALSE)
}
