% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{run.DOT.highresolution}
\alias{run.DOT.highresolution}
\title{A wrapper for running the DOT algorithm for high-resolution spatial data with suggested parameters}
\usage{
run.DOT.highresolution(object, lambda_a = 0, iterations = 100, verbose = FALSE)
}
\arguments{
\item{object}{A DOT object created using create.DOT().}

\item{lambda_a}{A value between 0 and 1 for matching ratio of cell types}

\item{iterations}{Integer. Maximum number of iterations of FW}

\item{verbose}{Boolean. Whether progress should be displayed.}
}
\value{
A DOT object with the produced results contained in the weights slot
}
\description{
A wrapper for running the DOT algorithm for high-resolution spatial data with suggested parameters
}
\examples{
data(dot.sample)
dot.ref <- setup.ref(dot.sample$ref$counts[, 1:1000], dot.sample$ref$labels[1:1000], 2)
dot.srt <- setup.srt(dot.sample$srt$counts, dot.sample$srt$coordinates)
dot <- create.DOT(dot.srt, dot.ref)
# No. iterations is reduced to 10 for this example (default is 100)
dot <- run.DOT.highresolution(dot, iterations = 10)
}
