% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{setup.ref}
\alias{setup.ref}
\title{Processing the reference single-cell data}
\usage{
setup.ref(
  ref_data,
  ref_annotations = NULL,
  ref_subcluster_size = 10,
  max_genes = 5000,
  remove_mt = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ref_data}{A gene x cell matrix of gene expressions. Can be a matrix-like object or a Seurat/AnnData object.}

\item{ref_annotations}{A character vector (one for each cell) or a single vector pointing to the slot in the Seurat/AnnData object}

\item{ref_subcluster_size}{An integer. Maximum number of sub-clusters per sub-population.}

\item{max_genes}{An integer. Maximum number of genes to pick.}

\item{remove_mt}{Boolean. Whether mitochondrial genes must be removed.}

\item{verbose}{Boolean. Whether progress should be displayed.}
}
\value{
A list containing the processed ref data.
}
\description{
Processing the reference single-cell data
}
\examples{
data(dot.sample)
dot.ref <- setup.ref(dot.sample$ref$counts[, 1:1000], dot.sample$ref$labels[1:1000], 2)
}
